logout

+++ src/script/login/login.js
@@ -56,6 +56,9 @@ async function validatelogin(col){
 	//let passpattern = new RegExp('^().{6,24}$');
 	var email = document.getElementById('email').value.toLowerCase();
 	var password = document.getElementById('password').value;
+	var keep = document.getElementById('keep').checked;
+	
+	console.log(keep);
 	
 	var username = String(email).split("@");
 	var testemail = username[0] + "@umkc.edu";
@@ -77,7 +80,10 @@ async function validatelogin(col){
 	
 	console.log(doc.exists);
 		if (doc.exists){
-			localStorage.setItem("ID", doc.id);
+			if(keep)
+				localStorage.setItem("ID", doc.id);
+			else 
+				sessionStorage.setItem("ID", doc.id);
 			result = true;
 		}
 	});
